python-service/requirements.txt
fastapi
uvicorn
pandas
plotly
scikit-learn
groq
python-multipart
kaleido
python-dotenv
